refactor(tools): close out the vault tools migration into src/tools/vault/ (#1310)
Move UpdateFrontmatterTool and AppendContentTool out of the flat
src/tools/vault-tools-extended.ts into the src/tools/vault/ package,
exported and re-exported like their nine siblings:

- src/tools/vault/update-frontmatter-tool.ts
- src/tools/vault/append-content-tool.ts

They join getVaultTools() appended last, preserving both tool ids
(update_frontmatter, append_content) and the registration order they
had under the separate 'vault-extended' source — tool ids are
model-facing and must not move. The now-empty vault-extended registry
source is removed. Pure move: no behaviour, signature, or naming
changes, and the buildDiffContext preamble clone the issue notes is
deliberately left for its own change.

Test tree mirrors the source layout: replace the two overlapping
vault-tools-extended test files (one flat, one in test/tools/vault/)
with per-tool files that keep every existing case, including the
buildDiffContext/applyConfirmedEdit coverage unique to the flat file
and the #910 wikilink-exclusion regressions.
