+++ test/services/tool-registrar.test.ts
@@ -2,11 +2,12 @@ import { ToolRegistrar } from '../../src/services/tool-registrar';
 
 // Mock all tool source modules
 vi.mock('../../src/tools/vault', () => ({
-	getVaultTools: () => [{ name: 'read_file' }, { name: 'write_file' }],
-}));
-
-vi.mock('../../src/tools/vault-tools-extended', () => ({
-	getExtendedVaultTools: () => [{ name: 'read_frontmatter' }],
+	getVaultTools: () => [
+		{ name: 'read_file' },
+		{ name: 'write_file' },
+		{ name: 'update_frontmatter' },
+		{ name: 'append_content' },
+	],
 }));
 
 vi.mock('../../src/tools/web-tools', () => ({
@@ -57,7 +58,8 @@ describe('ToolRegistrar', () => {
 
 			expect(mockRegistry.registerTool).toHaveBeenCalledWith(expect.objectContaining({ name: 'read_file' }));
 			expect(mockRegistry.registerTool).toHaveBeenCalledWith(expect.objectContaining({ name: 'write_file' }));
-			expect(mockRegistry.registerTool).toHaveBeenCalledWith(expect.objectContaining({ name: 'read_frontmatter' }));
+			expect(mockRegistry.registerTool).toHaveBeenCalledWith(expect.objectContaining({ name: 'update_frontmatter' }));
+			expect(mockRegistry.registerTool).toHaveBeenCalledWith(expect.objectContaining({ name: 'append_content' }));
 			expect(mockRegistry.registerTool).toHaveBeenCalledWith(expect.objectContaining({ name: 'google_search' }));
 			expect(mockRegistry.registerTool).toHaveBeenCalledWith(expect.objectContaining({ name: 'save_memory' }));
 			expect(mockRegistry.registerTool).toHaveBeenCalledWith(expect.objectContaining({ name: 'generate_image' }));
@@ -68,8 +70,8 @@ describe('ToolRegistrar', () => {
 		it('should register the correct total number of tools', async () => {
 			await registrar.registerAll(mockRegistry, mockLogger, mockPlugin);
 
-			// 2 vault + 1 extended + 1 web + 1 memory + 1 image + 1 skill + 1 session-recall = 8
-			expect(mockRegistry.registerTool).toHaveBeenCalledTimes(8);
+			// 4 vault + 1 web + 1 memory + 1 image + 1 skill + 1 session-recall = 9
+			expect(mockRegistry.registerTool).toHaveBeenCalledTimes(9);
 		});
 
 		it('should skip Gemini-only sources when provider is ollama', async () => {
@@ -117,15 +119,15 @@ describe('ToolRegistrar', () => {
 		it('should continue registering other sources if one fails', async () => {
 			// Make registerTool throw for a specific tool
 			mockRegistry.registerTool.mockImplementation((tool: any) => {
-				if (tool.name === 'read_frontmatter') {
+				if (tool.name === 'write_file') {
 					throw new Error('Registration failed');
 				}
 			});
 
 			await registrar.registerAll(mockRegistry, mockLogger, mockPlugin);
 
-			// Should log the error for the vault-extended source
-			expect(mockLogger.error).toHaveBeenCalledWith('Failed to register vault-extended tools:', expect.any(Error));
+			// Should log the error for the vault source
+			expect(mockLogger.error).toHaveBeenCalledWith('Failed to register vault tools:', expect.any(Error));
 			// Should still have registered tools from other sources
 			expect(mockRegistry.registerTool).toHaveBeenCalledWith(expect.objectContaining({ name: 'read_file' }));
 			expect(mockRegistry.registerTool).toHaveBeenCalledWith(expect.objectContaining({ name: 'google_search' }));
@@ -139,7 +141,8 @@ describe('ToolRegistrar', () => {
 
 			expect(mockRegistry.unregisterTool).toHaveBeenCalledWith('read_file');
 			expect(mockRegistry.unregisterTool).toHaveBeenCalledWith('write_file');
-			expect(mockRegistry.unregisterTool).toHaveBeenCalledWith('read_frontmatter');
+			expect(mockRegistry.unregisterTool).toHaveBeenCalledWith('update_frontmatter');
+			expect(mockRegistry.unregisterTool).toHaveBeenCalledWith('append_content');
 			expect(mockRegistry.unregisterTool).toHaveBeenCalledWith('google_search');
 			expect(mockRegistry.unregisterTool).toHaveBeenCalledWith('save_memory');
 			expect(mockRegistry.unregisterTool).toHaveBeenCalledWith('generate_image');
@@ -149,15 +152,15 @@ describe('ToolRegistrar', () => {
 
 		it('should continue unregistering other sources if one fails', async () => {
 			mockRegistry.unregisterTool.mockImplementation((name: string) => {
-				if (name === 'read_frontmatter') {
+				if (name === 'write_file') {
 					throw new Error('Unregistration failed');
 				}
 			});
 
 			await registrar.unregisterAll(mockRegistry, mockLogger);
 
-			// Should log debug for the vault-extended source failure
-			expect(mockLogger.debug).toHaveBeenCalledWith('Failed to unregister vault-extended tools:', expect.any(Error));
+			// Should log debug for the vault source failure
+			expect(mockLogger.debug).toHaveBeenCalledWith('Failed to unregister vault tools:', expect.any(Error));
 			// Should still have unregistered tools from other sources
 			expect(mockRegistry.unregisterTool).toHaveBeenCalledWith('read_file');
 			expect(mockRegistry.unregisterTool).toHaveBeenCalledWith('google_search');
